+++ ModbusTCP/View/CommunicationView.xaml
@@ -15,7 +15,8 @@
         </Grid.RowDefinitions>
         <DockPanel Grid.Row="0" Margin="0 0 0 0">
             <Button Margin="20 0 0 0" Content="START" x:Name="StartButton" Command="{Binding StartCommand}" Height="30" HorizontalAlignment="Left" Width="100" FontSize="16"/>
-            <Button Margin="20,35,0,35" Content="PAUSE" x:Name="PauseButton" Command="{Binding PauseCommand}" Height="30" HorizontalAlignment="Left" Width="100" FontSize="16"/>
+            <Button Margin="20 0 0 0" x:Name="Connect" Command="{Binding ConnectCommand}" Content="CONNECT" HorizontalAlignment="Left" Height="28" VerticalAlignment="Center" Width="90" FontSize="16"/>
+            <Button Margin="20 0 0 0" x:Name="Disconnect" Command="{Binding DisconnectCommand}" Content="DISCONNECT" HorizontalAlignment="Left" Height="28" VerticalAlignment="Center" Width="110" FontSize="16"/>
         </DockPanel>
         <Grid Grid.Row="1">
             <local:ListBoxScroll Margin="10 10 10 10" VerticalAlignment="Top"
